Remove app source & update to match main site
The "app-source" folder and its contents were removed from this fork as they didn't need to be there given how this version of the Playground is only meant to provide the Playground experience, but it's not going to be used for further developing it.

The "app-source" and all other materials are still available on the main site: www.solar2dplayground.com.